remove the double shader compile
